ingest/requirements.txt
elasticsearch==5.4.0
click==6.7
uhashring==0.7
click-configfile==0.2.2
retrying==1.3.3
PyYaml==3.12
psutil==5.4.1
